tests/requirements-test.txt
pytest>=7.0.0
pytest-mock>=3.10.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
nbformat>=5.7.0
nbclient>=0.7.0
ipython>=8.0.0
pyzmq>=24.0.0 